requirements.txt
streamlit
openai
pydub
gtts
faiss-cpu
numpy
pandas
python-dotenv